Fixes #2172, don't allow persist etc if no tx

+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/entitymanager/TransactionScopedEntityManager.java
@@ -28,6 +28,7 @@ import javax.persistence.FlushModeType;
 import javax.persistence.LockModeType;
 import javax.persistence.Query;
 import javax.persistence.StoredProcedureQuery;
+import javax.persistence.TransactionRequiredException;
 import javax.persistence.TypedQuery;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaDelete;
@@ -43,6 +44,7 @@ import io.quarkus.hibernate.orm.runtime.RequestScopedEntityManagerHolder;
 
 public class TransactionScopedEntityManager implements EntityManager {
 
+    protected static final String TRANSACTION_IS_NOT_ACTIVE = "Transaction is not active, consider adding @Transactional to your method to automatically activate one.";
     private final TransactionManager transactionManager;
     private final TransactionSynchronizationRegistry tsr;
     private final EntityManagerFactory emf;
@@ -65,7 +67,7 @@ public class TransactionScopedEntityManager implements EntityManager {
         if (isInTransaction()) {
             EntityManager em = (EntityManager) tsr.getResource(transactionKey);
             if (em != null) {
-                return new EntityManagerResult(em, false);
+                return new EntityManagerResult(em, false, true);
             }
             EntityManager newEm = emf.createEntityManager();
             newEm.joinTransaction();
@@ -82,13 +84,13 @@ public class TransactionScopedEntityManager implements EntityManager {
                     newEm.close();
                 }
             });
-            return new EntityManagerResult(newEm, false);
+            return new EntityManagerResult(newEm, false, true);
         } else {
             //this will throw an exception if the request scope is not active
             //this is expected as either the request scope or an active transaction
             //is required to properly managed the EM lifecycle
             RequestScopedEntityManagerHolder requestScopedEms = this.requestScopedEms.get();
-            return new EntityManagerResult(requestScopedEms.getOrCreateEntityManager(unitName, emf), false);
+            return new EntityManagerResult(requestScopedEms.getOrCreateEntityManager(unitName, emf), false, false);
         }
     }
 
@@ -112,6 +114,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void persist(Object entity) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.persist(entity);
         }
     }
@@ -119,6 +124,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public <T> T merge(T entity) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             return emr.em.merge(entity);
         }
     }
@@ -126,6 +134,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void remove(Object entity) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.remove(entity);
         }
     }
@@ -168,6 +179,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void flush() {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.flush();
         }
     }
@@ -189,6 +203,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void lock(Object entity, LockModeType lockMode) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.lock(entity, lockMode);
         }
     }
@@ -196,6 +213,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void lock(Object entity, LockModeType lockMode, Map<String, Object> properties) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.lock(entity, lockMode, properties);
         }
     }
@@ -203,6 +223,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void refresh(Object entity) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.refresh(entity);
         }
     }
@@ -210,6 +233,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void refresh(Object entity, Map<String, Object> properties) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.refresh(entity, properties);
         }
     }
@@ -217,6 +243,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void refresh(Object entity, LockModeType lockMode) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.refresh(entity, lockMode);
         }
     }
@@ -224,6 +253,9 @@ public class TransactionScopedEntityManager implements EntityManager {
     @Override
     public void refresh(Object entity, LockModeType lockMode, Map<String, Object> properties) {
         try (EntityManagerResult emr = getEntityManager()) {
+            if (!emr.allowModification) {
+                throw new TransactionRequiredException(TRANSACTION_IS_NOT_ACTIVE);
+            }
             emr.em.refresh(entity, lockMode, properties);
         }
     }
@@ -463,12 +495,14 @@ public class TransactionScopedEntityManager implements EntityManager {
 
     static class EntityManagerResult implements AutoCloseable {
 
-        private final EntityManager em;
-        private final boolean closeOnEnd;
+        final EntityManager em;
+        final boolean closeOnEnd;
+        final boolean allowModification;
 
-        EntityManagerResult(EntityManager em, boolean closeOnEnd) {
+        EntityManagerResult(EntityManager em, boolean closeOnEnd, boolean allowModification) {
             this.em = em;
             this.closeOnEnd = closeOnEnd;
+            this.allowModification = allowModification;
         }
 
         @Override
